zain-backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
